+++ src/main/kotlin/com/dash/repository/WidgetRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface WidgetRepository : JpaRepository<Widget, Long> {
+interface WidgetRepository : JpaRepository<Widget, Int> {
     fun findByTabIdOrderByWidgetOrderAsc(tabId: Int): List<Widget>
 }
